+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Improved extendability: one may override `Magic::Decoratable#decorator_base` to be used for lookups.
+- `Magic::Decoratable.classes` for all the decoratables.
 
 
 ## [0.2.0] — 2024-10-17